AB-35: Extract common types to generic folder.
Use primitive import feature for the time being.
